+++ src/Init.php
@@ -20,9 +20,9 @@ if ( ! defined( 'ABSPATH' ) ) {
 class Init {
 
 	/**
-	 * Universal loader.
+	 * Universal settings/updater loader.
 	 *
-	 * Load the correct updater from a single function.
+	 * Load the correct settings/updater from a single function.
 	 *
 	 * @param array $config Configuration data for plugin or theme.
 	 *
@@ -35,6 +35,24 @@ class Init {
 		}
 		if ( in_array( 'theme', $config, true ) ) {
 			( new Theme_Updater_Admin( $config ) )->load_hooks();
+		}
+	}
+
+	/**
+	 * Universal updater.
+	 *
+	 * Load the correct updater from a single function.
+	 *
+	 * @param array $config Configuration data for plugin or theme.
+	 *
+	 * @return void
+	 */
+	public function updater( $config ) {
+		if ( in_array( 'plugin', $config, true ) ) {
+			( new Plugin_Updater_Admin( $config ) )->updater();
+		}
+		if ( in_array( 'theme', $config, true ) ) {
+			( new Theme_Updater_Admin( $config ) )->updater();
 		}
 	}
 }